feat: slight improvment to user experience
